chore: PR feedback pt 1
